+++ src/main/java/potato/DateTime.java
@@ -1,3 +1,5 @@
+package potato;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 public class DateTime {
