+++ src/main/java/potato/Parser.java
@@ -1,3 +1,6 @@
+package potato;
+
+import potato.command.*;
 public class Parser {
     public static Command parse(String input) {
         if (input.equals("bye")) {
