+++ src/main/java/potato/PotatoException.java
@@ -1,3 +1,5 @@
+package potato;
+
 public class PotatoException extends Throwable {
     protected String message;
     public PotatoException(String message) {
