+++ src/main/java/potato/Storage.java
@@ -1,3 +1,6 @@
+package potato;
+
+import potato.task.*;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
