+++ src/main/java/potato/command/AddCommand.java
@@ -1,13 +1,16 @@
+package potato.command;
+
+import potato.*;
 import java.io.IOException;
 
-public class MarkCommand extends Command{
-    public MarkCommand(String input) {
+public class AddCommand extends Command{
+    public AddCommand(String input) {
         super.isExit = false;
         super.input = input;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws IOException {
-        tasks.mark(input, storage);
+        tasks.add(input, storage);
     }
 }
