+++ src/main/java/potato/command/ExitCommand.java
@@ -1,3 +1,7 @@
+package potato.command;
+
+import potato.*;
+
 public class ExitCommand extends Command{
     public ExitCommand(String input) {
         super.isExit = true;
