+++ src/main/java/potato/command/UnmarkCommand.java
@@ -1,13 +1,16 @@
+package potato.command;
+
+import potato.*;
 import java.io.IOException;
 
-public class AddCommand extends Command{
-    public AddCommand(String input) {
+public class UnmarkCommand extends Command{
+    public UnmarkCommand(String input) {
         super.isExit = false;
         super.input = input;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws IOException {
-        tasks.add(input, storage);
+        tasks.unmark(input, storage);
     }
 }
