+++ src/main/java/potato/task/Task.java
@@ -1,3 +1,5 @@
+package potato.task;
+
 public class Task {
     protected String description;
     protected boolean isDone;
